# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  physical: "physical",
  digital: "digital",
  curriculum: "curriculum",
} as const;

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: ProductCategory;
  imageUrl: string;
  gumroadUrl?: string | null;
  downloadUrl?: string | null;
  featured: boolean;
  inStock: boolean;
  createdAt: string;
}

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  description?: string | null;
  createdAt: string;
}

export interface Tutorial {
  id: number;
  title: string;
  youtubeId: string;
  description?: string | null;
  featured: boolean;
  createdAt: string;
}

export interface Download {
  id: number;
  title: string;
  description?: string | null;
  fileUrl: string;
  fileType: string;
  thumbnailUrl?: string | null;
  createdAt: string;
}

export interface CheckoutRequest {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CheckoutResponse {
  url: string;
  sessionId: string;
}

export interface OrderSuccess {
  productName: string;
  productCategory: string;
  gumroadUrl?: string | null;
  downloadUrl?: string | null;
  email?: string | null;
}

export interface ContactRequest {
  name: string;
  email: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
  category?: ListProductsCategory;
};

export type ListProductsCategory =
  (typeof ListProductsCategory)[keyof typeof ListProductsCategory];

export const ListProductsCategory = {
  physical: "physical",
  digital: "digital",
  curriculum: "curriculum",
} as const;

export type ListTutorialsParams = {
  featured?: boolean;
};

export type GetOrderSuccessParams = {
  session_id: string;
};