# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blade & Quill Art Academy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: gallery
    description: Gallery artwork
  - name: tutorials
    description: Tutorial videos
  - name: downloads
    description: Free downloads
  - name: checkout
    description: Stripe checkout
  - name: contact
    description: Contact form
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [physical, digital, curriculum]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery artworks
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"

  /tutorials:
    get:
      operationId: listTutorials
      tags: [tutorials]
      summary: List tutorial videos
      parameters:
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Tutorial list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tutorial"

  /downloads:
    get:
      operationId: listDownloads
      tags: [downloads]
      summary: List free downloadable resources
      responses:
        "200":
          description: Downloads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"

  /checkout:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /checkout/success:
    get:
      operationId: getOrderSuccess
      tags: [checkout]
      summary: Get order success info after Stripe redirect
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order success details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSuccess"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
      responses:
        "200":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
          enum: [physical, digital, curriculum]
        imageUrl:
          type: string
        gumroadUrl:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        inStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, price, category, imageUrl, featured, inStock, createdAt]

    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, imageUrl, createdAt]

    Tutorial:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        youtubeId:
          type: string
        description:
          type: string
          nullable: true
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, youtubeId, featured, createdAt]

    Download:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        fileUrl:
          type: string
        fileType:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, fileUrl, fileType, createdAt]

    CheckoutRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
      required: [productId, quantity]

    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required: [url, sessionId]

    OrderSuccess:
      type: object
      properties:
        productName:
          type: string
        productCategory:
          type: string
        gumroadUrl:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required: [productName, productCategory]

    ContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        message:
          type: string
      required: [name, email, message]

    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]