# lib/api-zod/src/generated/types/checkoutResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutResponse {
  url: string;
  sessionId: string;
}