# lib/api-zod/src/generated/types/download.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Download {
  id: number;
  title: string;
  description?: string | null;
  fileUrl: string;
  fileType: string;
  thumbnailUrl?: string | null;
  createdAt: Date;
}