# lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  description?: string | null;
  createdAt: Date;
}