# lib/api-zod/src/generated/types/orderSuccess.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderSuccess {
  productName: string;
  productCategory: string;
  gumroadUrl?: string | null;
  downloadUrl?: string | null;
  email?: string | null;
}