# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: ProductCategory;
  imageUrl: string;
  gumroadUrl?: string | null;
  downloadUrl?: string | null;
  featured: boolean;
  inStock: boolean;
  createdAt: Date;
}