# lib/api-zod/src/generated/types/tutorial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blade & Quill Art Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Tutorial {
  id: number;
  title: string;
  youtubeId: string;
  description?: string | null;
  featured: boolean;
  createdAt: Date;
}